backend/requirements.txt
fastapi==0.115.3
uvicorn==0.31.0
python-dotenv==1.0.0
groq==0.9.0
httpx==0.27.2
pydantic>=2.5.2
